+++ libs/libc/termios/termios.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <sys/ioctl.h>
 #include <termios.h>
 
@@ -13,19 +14,19 @@ int tcsetpgrp(int fd, pid_t pgid)
 
 int tcgetattr(int fd, termios_t* termios_p)
 {
-    return ioctl(0, TCGETS, (int)termios_p);
+    return ioctl(0, TCGETS, (uintptr_t)termios_p);
 }
 
 int tcsetattr(int fd, int optional_actions, const termios_t* termios_p)
 {
     if (optional_actions == TCSANOW) {
-        return ioctl(0, TCSETS, (int)termios_p);
+        return ioctl(0, TCSETS, (uintptr_t)termios_p);
     }
     if (optional_actions == TCSADRAIN) {
-        return ioctl(0, TCSETSW, (int)termios_p);
+        return ioctl(0, TCSETSW, (uintptr_t)termios_p);
     }
     if (optional_actions == TCSAFLUSH) {
-        return ioctl(0, TCSETSF, (int)termios_p);
+        return ioctl(0, TCSETSF, (uintptr_t)termios_p);
     }
     return -1;
 }
